improve util accessSpreadsheet
