+++ nodes/BorderHealthInsuranceSwitzerland/utils.ts
@@ -1,27 +1,4 @@
-import { JWT } from "google-auth-library";
-import { GoogleSpreadsheet, GoogleSpreadsheetRow } from "google-spreadsheet";
-
-export async function accessSpreadsheet() {
-  const credentials = JSON.parse(
-    process.env.SPREADSHEET_API_CREDENTIALS ?? ""
-  ) as Credentials;
-  const serviceAccountAuth = new JWT({
-    email: credentials?.client_email,
-    key: credentials?.private_key,
-    scopes: [
-      "https://www.googleapis.com/auth/spreadsheets",
-      "https://www.googleapis.com/auth/drive.file",
-    ],
-  });
-
-  const doc = new GoogleSpreadsheet(
-    "1QbuYpRlCEk37o1nYc08rX2Na2OM3rXac6jfaQSi8sWU",
-    serviceAccountAuth
-  );
-
-  await doc.loadInfo(); // loads document properties and worksheets
-  return doc;
-}
+import { GoogleSpreadsheetRow } from "google-spreadsheet";
 
 export function replaceSpecialWithNormal(input: string) {
   // Normalize the string to separate base letters from diacritics
